fix share reading

+++ include/xuanwu/DevicesGroup.h
@@ -57,7 +57,7 @@ namespace Xuanwu {
         auto operator()(const TInjector &injector, const TDependency &dep) const {
             CLOG(DEBUG, "Device") << "MultipleDeviceGroup run injector With " << num_ << " " << typeid(Device).name();
             auto ret = static_cast<const MultipleDevicesGroup<OtherDevices...> *>(this)->operator()(injector, dep);
-            for (size_t i = 0; i < num_; i++) {
+            for (int i = 0; i < num_; i++) {
                 CLOG(DEBUG, "Device") << "MyDeviceGroup creating CPUDevice " << i;
                 ret->push_back(injector.template create<std::unique_ptr<Device>>());
                 CLOG(DEBUG, "Device") << "MyDeviceGroup creating CPUDevice " << i << " OK";
